+++ NEWS.md
@@ -30,6 +30,8 @@ Bug fixes and minor improvements
   It also returns `TRUE` for `svglite::svglite()` if `packageVersion("svglite") >= '2.1.0'`.
 * `clippingPathGrob()` will now consistently - as previously documented - use `ragg::agg_png()` 
   as the default `png_device` graphics device if it is available and `use_R4.1_clipping` is `FALSE`.
+* `grid.pattern_image()` should no longer throw an inscrutable `Error in magick_image_readpath`
+  error on certain platforms such as Windows (#47).
 
 gridpattern v0.4.0
 ==================